Only push rpms that have changed
This is our first step towards making this a serious repo updater.
Currently, pushing RPMS every time generates different packageids,
meaning we have to manually clean the metadata for our repo. Eventually,
we'll never need to do this, and I'd like to get to that point as
quickly as possible.