Only push new packages
This is getting to a point where I like it. It doesn't corrupt the
repository since it's very conservative about replacing packages. I
originally wanted to detect if a package had changed, but I think I'll
forgo that: it's already a difficult task and breaking the repository
should be "hard" to do. Changes in a local repository will cause new
packages to be made, but there's no reason why they should be
automatically pushed to the server.