Serious refactoring
dafrito-yum is now a command hub, for pushing, pulling, and building.
This gives me a lot more flexibility in how I manage and test my
repository.

While it's designed with genericity in mind, it's very specific to my
setup, so modifications to the config are needed to get it to work.